requirements.txt
Django>=1.8.11,<1.9a
argparse>=1.2.1
django-registration==2.0.4
django-tinymce==1.5.3
ipaddr==2.1.11
python-social-auth==0.2.14
requests==2.7.0
wsgiref==0.1.2
lxml>=2.3.2
PyYAML>=3.10,<4.0a
djangorestframework==3.5.3